requirements.txt
uagents
httpx
pydantic
python-dotenv
youtube-search-python
requests
aiohttp
#hyperon>=0.2.8